+++ wow_world_base/src/inner/shared/item_class_vanilla_tbc.rs
@@ -115,7 +115,7 @@ impl TryFrom<u8> for ItemClass {
             13 => Ok(Self::Key),
             14 => Ok(Self::Permanent),
             15 => Ok(Self::Misc),
-            v => Err(crate::errors::EnumError::new("ItemClass", v as u32),)
+            v => Err(crate::errors::EnumError::new("ItemClass", v as u64),)
         }
     }
 }
