+++ wow_world_base/src/inner/shared/item_quality_vanilla_tbc.rs
@@ -70,7 +70,7 @@ impl TryFrom<u8> for ItemQuality {
             4 => Ok(Self::Epic),
             5 => Ok(Self::Legendary),
             6 => Ok(Self::Artifact),
-            v => Err(crate::errors::EnumError::new("ItemQuality", v as u32),)
+            v => Err(crate::errors::EnumError::new("ItemQuality", v as u64),)
         }
     }
 }
