+++ wow_world_base/src/inner/tbc/creature_family.rs
@@ -195,7 +195,7 @@ impl TryFrom<u8> for CreatureFamily {
             34 => Ok(Self::NetherRay),
             35 => Ok(Self::Serpent),
             36 => Ok(Self::SeaLion),
-            v => Err(crate::errors::EnumError::new("CreatureFamily", v as u32),)
+            v => Err(crate::errors::EnumError::new("CreatureFamily", v as u64),)
         }
     }
 }
