+++ wow_world_base/src/inner/vanilla/skill.rs
@@ -655,7 +655,7 @@ impl TryFrom<u16> for Skill {
             754 => Ok(Self::RacialHuman),
             758 => Ok(Self::PetEventRc),
             762 => Ok(Self::Riding),
-            v => Err(crate::errors::EnumError::new("Skill", v as u32),)
+            v => Err(crate::errors::EnumError::new("Skill", v as u64),)
         }
     }
 }
