+++ wow_world_base/src/inner/wrath/item_quality.rs
@@ -75,7 +75,7 @@ impl TryFrom<u8> for ItemQuality {
             5 => Ok(Self::Legendary),
             6 => Ok(Self::Artifact),
             7 => Ok(Self::Heirloom),
-            v => Err(crate::errors::EnumError::new("ItemQuality", v as u32),)
+            v => Err(crate::errors::EnumError::new("ItemQuality", v as u64),)
         }
     }
 }
