+++ wow_world_messages/src/world/shared/arena_team_command_tbc_wrath.rs
@@ -54,7 +54,7 @@ impl TryFrom<u32> for ArenaTeamCommand {
             1 => Ok(Self::TeamInviteSs),
             3 => Ok(Self::TeamQuitS),
             14 => Ok(Self::TeamFounderS),
-            v => Err(crate::errors::EnumError::new("ArenaTeamCommand", v),)
+            v => Err(crate::errors::EnumError::new("ArenaTeamCommand", v as u64),)
         }
     }
 }
