+++ wow_world_messages/src/world/shared/corpse_query_result_vanilla_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u8> for CorpseQueryResult {
         match value {
             0 => Ok(Self::NotFound),
             1 => Ok(Self::Found),
-            v => Err(crate::errors::EnumError::new("CorpseQueryResult", v as u32),)
+            v => Err(crate::errors::EnumError::new("CorpseQueryResult", v as u64),)
         }
     }
 }
