+++ wow_world_messages/src/world/shared/declined_pet_name_included_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u8> for DeclinedPetNameIncluded {
         match value {
             0 => Ok(Self::NotIncluded),
             1 => Ok(Self::Included),
-            v => Err(crate::errors::EnumError::new("DeclinedPetNameIncluded", v as u32),)
+            v => Err(crate::errors::EnumError::new("DeclinedPetNameIncluded", v as u64),)
         }
     }
 }
