+++ wow_world_messages/src/world/shared/dungeon_difficulty_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u8> for DungeonDifficulty {
         match value {
             0 => Ok(Self::Normal),
             1 => Ok(Self::Heroic),
-            v => Err(crate::errors::EnumError::new("DungeonDifficulty", v as u32),)
+            v => Err(crate::errors::EnumError::new("DungeonDifficulty", v as u64),)
         }
     }
 }
