+++ wow_world_messages/src/world/shared/gm_ticket_status_response_vanilla_tbc_wrath.rs
@@ -49,7 +49,7 @@ impl TryFrom<u32> for GmTicketStatusResponse {
             1 => Ok(Self::Updated),
             2 => Ok(Self::Closed),
             3 => Ok(Self::Survey),
-            v => Err(crate::errors::EnumError::new("GmTicketStatusResponse", v),)
+            v => Err(crate::errors::EnumError::new("GmTicketStatusResponse", v as u64),)
         }
     }
 }
