+++ wow_world_messages/src/world/shared/loot_method_vanilla_tbc_wrath.rs
@@ -82,7 +82,7 @@ impl TryFrom<u8> for LootMethod {
             20 => Ok(Self::Fishinghole),
             21 => Ok(Self::FishingFail),
             22 => Ok(Self::Insignia),
-            v => Err(crate::errors::EnumError::new("LootMethod", v as u32),)
+            v => Err(crate::errors::EnumError::new("LootMethod", v as u64),)
         }
     }
 }
